+++ prefect_email/__init__.py
@@ -1,5 +1,5 @@
 from . import _version
-from .credentials import SMTPType, SMTPServer, EmailCredentials  # noqa
+from .credentials import SMTPType, SMTPServer, EmailServerCredentials  # noqa
 from .message import email_send_message  # noqa
 
 __version__ = _version.get_versions()["version"]